requirements-rag.txt
# MATLAB RAG Assistant - Python Dependencies

# Core FastAPI & Uvicorn
fastapi
uvicorn
python-multipart

# Streamlit Frontend
streamlit==1.28.1
requests==2.31.0

# LangChain & LLM Integration
langchain==0.0.350
langchain-core==0.0.11
langchain-openai==0.0.8
langchain-anthropic==0.1.5
langchain-text-splitters==0.0.1

# Vector Database
pinecone-client==3.0.0

# API & Async
aiohttp==3.9.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Optional: Production Server
gunicorn==21.2.0

# Optional: Development Tools
pytest==7.4.3
black==23.12.0
flake8==6.1.0
